requirements.txt
asgiref==3.6.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.0.6
django-cors-headers==3.14.0
django-rest-registration==0.7.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
filelock==3.9.0
Flask==2.2.2
Flask-CLI==0.4.0
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.4.0
platformdirs==2.6.2
Pygments==2.14.0
PyJWT==1.7.1
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.3
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
virtualenv==20.17.1
Werkzeug==2.2.2